Rename and change routes to be not be design specific
