+++ router/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
 module.exports = function(app) {
-  app.use('/new-design-applicant', require('./routes/new_design_applicant'));
-  app.use('/status-design-applicant', require('./routes/status_design_applicant'));
+  app.use('/new-applicant', require('./routes/new_applicant'));
+  app.use('/status-applicant', require('./routes/status_applicant'));
 };
